requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0
opencv-python>=4.7.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Progress tracking
tqdm>=4.65.0
rich>=13.3.0

# Model evaluation
torchmetrics>=0.11.0
lpips>=0.1.4
scikit-image>=0.20.0

# Web interface
gradio>=3.40.0
fastapi>=0.95.0
uvicorn>=0.22.0

# Logging and experiments
wandb>=0.15.0
tensorboard>=2.13.0

# Code quality
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
pytest>=7.3.0

# Utilities
einops>=0.6.1
accelerate>=0.20.0
omegaconf>=2.3.0